# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/purchase_receipt_items:
    get:
      tags:
        - PurchaseReceiptItems
      summary: List PurchaseReceiptItems
      description: Get a paginated list of PurchaseReceiptItems with optional filtering and sorting
      operationId: listPurchaseReceiptItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReceiptItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseReceiptItems
      summary: Create PurchaseReceiptItem
      description: Create a new PurchaseReceiptItem
      operationId: createPurchaseReceiptItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseReceiptItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReceiptItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/stock_entries:
    get:
      tags:
        - StockEntries
      summary: List StockEntries
      description: Get a paginated list of StockEntries with optional filtering and sorting
      operationId: listStockEntries
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockEntryList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - StockEntries
      summary: Create StockEntry
      description: Create a new StockEntry
      operationId: createStockEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PurchaseReceiptItem:
      type: object
      required:
        - id
        - receipt_id
        - company_id
        - item_id
        - quantity
        - rate
        - amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        receipt_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        rate:
          type: string
        amount:
          type: string
        metadata:
          type: object
...
    CreatePurchaseReceiptItemRequest:
      type: object
      required:
        - receipt_id
        - company_id
        - item_id
        - quantity
        - rate
        - amount
        - metadata
      properties:
        receipt_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        quantity:
          type: string
        rate:
          type: string
        amount:
          type: string
        metadata:
          type: object
...
    PurchaseReceiptItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseReceiptItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    StockEntry:
      type: object
      required:
        - id
        - entry_number
        - company_id
        - stock_entry_type
        - posting_date
        - status
        - posting_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        entry_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        stock_entry_type:
          type: string
        from_warehouse_id:
          nullable: true
          type: string
          format: uuid
        to_warehouse_id:
          nullable: true
          type: string
          format: uuid
        posting_date:
          type: string
          format: date
        status:
          type: string
        posting_state:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreateStockEntryRequest:
      type: object
      required:
        - entry_number
        - company_id
        - stock_entry_type
        - posting_date
        - status
        - posting_state
        - metadata
      properties:
        entry_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        stock_entry_type:
          type: string
        from_warehouse_id:
          nullable: true
          type: string
          format: uuid
        to_warehouse_id:
          nullable: true
          type: string
          format: uuid
        posting_date:
          type: string
          format: date
        status:
          type: string
        posting_state:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    StockEntryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockEntry'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []